requirements.txt
google-genai
pydantic
python-dotenv
pyyaml
pytest
